// packages/apply-rulesets/src/types.ts
export type CloudscapeRepo =
  | "actions"
  | "board-components"
  | "browser-test-tools"
  | "build-tools"
  | "chart-components"
  | "chat-components"
  | "code-view"
  | "collection-hooks"
  | "component-toolkit"
  | "components"
  | "demos"
  | "documenter"
  | "global-styles"
  | "jest-preset"
  | "test-utils"
  | "theming-core"